requirements.txt
transformers
torch
whisper
streamlit
streamlit-audiorecorder
apiclient
httplib2
oauth2client
bardapi
ffmpeg
ffmpeg-python
google-api-python-client
google-auth
openai-whisper
google-auth-oauthlib 
google-auth-httplib2
sentencepiece
huggingface_hub
datasets